fix: close three small debts in one pass
- `productos` detail and edit URLs asked for `(?P<pk>\d+)` while `Producto.codigo`
  and `GrupoProductos.codigo` are CharField primary keys, so the product list
  answered 500 as soon as a code was not numeric. The two `UnidadMedida` patterns
  keep `\d+`, because their pk is an automatic integer.

- Seven menu sections opened an empty submenu: two in `administracion`, two in
  `contabilidad`, two in `productos` and one in `requerimientos`.

- `cantidad`, `precio` and `valor` were `IntegerField` in ten places while the
  models declare `DecimalField` with 8 or 5 decimals, and the widgets already
  carried the `decimal` class. They now match the field they feed.

+++ almacen/forms.py
@@ -49,10 +49,11 @@ class FormularioDetalleMovimiento(forms.Form):
     nombre = forms.CharField(max_length=100, widget=forms.TextInput(attrs={'size': 35, 'class': 'form-control'}))
     unidad = forms.CharField(max_length=6,
                              widget=forms.TextInput(attrs={'size': 6, 'readonly': "readonly", 'class': 'form-control'}))
-    cantidad = forms.IntegerField(widget=forms.TextInput(attrs={'size': 6, 'class': 'form-control decimal'}))
-    precio = forms.IntegerField(widget=forms.TextInput(
+    cantidad = forms.DecimalField(max_digits=25, decimal_places=8,
+                                  widget=forms.TextInput(attrs={'size': 6, 'class': 'form-control decimal'}))
+    precio = forms.DecimalField(max_digits=25, decimal_places=8, widget=forms.TextInput(
         attrs={'size': 7, 'readonly': "readonly", 'class': 'form-control decimal'}))
-    valor = forms.IntegerField(widget=forms.TextInput(
+    valor = forms.DecimalField(max_digits=25, decimal_places=8, widget=forms.TextInput(
         attrs={'size': 10, 'readonly': "readonly", 'class': 'form-control decimal'}))
